Actually fix the conference begin/end time for real
Those date fields need to either be all quoted or not quoted,
otherwise some of them will be parsed as strings and others are dates,
and that breaks at least the code of _pages/calendar-archive.md.

The initial fix attempt restored the quotes, but Netlify CMS
now correctly saves them without quotes, and this led
to the initial breakage (the old entries didn't have quotes).
Then just removes all quotes.